requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
